+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v8.3.0 - ?
+## v8.3.0 - 2025-03-12
 
 - Add std::format plugin to enjoy advanced python string formatting in the inmanta DSL
 - Added create_environment_reference plugin to create reference to environment variables